Added new tables

+++ source/org.ohdsi.cdm.framework.etl/org.ohdsi.cdm.framework.etl.common/ETL/Common/Scripts/MySql/v5.3/CreateTables.sql
@@ -43,7 +43,7 @@ GO
 
 CREATE TABLE {sc}.cohort
 (
-   cohort_definition_id  int(11)   NOT NULL,
+   cohort_definition_id  bigint(20)   NOT NULL,
    subject_id            int(11)   NOT NULL,
    cohort_start_date     date      NOT NULL,
    cohort_end_date       date      NOT NULL
@@ -52,7 +52,7 @@ GO
 
 CREATE TABLE {sc}.cohort_attribute
 (
-   cohort_definition_id     int(11)   NOT NULL,
+   cohort_definition_id     bigint(20)   NOT NULL,
    subject_id               int(11)   NOT NULL,
    cohort_start_date        date      NOT NULL,
    cohort_end_date          date      NOT NULL,
@@ -64,7 +64,7 @@ GO
 
 CREATE TABLE {sc}.cohort_definition
 (
-   cohort_definition_id           int(11)        NOT NULL,
+   cohort_definition_id           bigint(20)        NOT NULL,
    cohort_definition_name         varchar(255)   NOT NULL,
    cohort_definition_description  text,
    definition_type_concept_id     int(11)        NOT NULL,
