+++ source/org.ohdsi.cdm.framework.etl/org.ohdsi.cdm.framework.etl.common/ETL/Common/Scripts/MySql/v5.4/CreateTables.sql
@@ -475,7 +475,7 @@ GO
 
 
 CREATE TABLE {sc}.cohort (
-			cohort_definition_id int(11) NOT NULL,
+			cohort_definition_id bigint(20) NOT NULL,
 			subject_id int(11) NOT NULL,
 			cohort_start_date date NOT NULL,
 			cohort_end_date date NOT NULL )ENGINE=InnoDB;
@@ -483,7 +483,7 @@ GO
 
 
 CREATE TABLE {sc}.cohort_definition (
-			cohort_definition_id int(11) NOT NULL,
+			cohort_definition_id bigint(20) NOT NULL,
 			cohort_definition_name varchar(255) NOT NULL,
 			cohort_definition_description text NULL,
 			definition_type_concept_id int(11) NOT NULL,
